requirements.txt
fastapi>=0.129,<1.0
uvicorn>=0.41,<1.0
gradio>=6.6,<7.0
openai>=2.20,<3.0
python-dotenv>=1.2,<2.0
psycopg2-binary>=2.9,<3.0
tiktoken>=0.12,<1.0
